Mapping JSON path (query) to K8S objects
This is a kubectl plugin to query values from all (within namespace or all) objects of a kind. The query expression is JSON path. I
planned YAML path but could not find a existing crate to parse it.

`kubectl get --filter` only supports a few items, and varies amount different resources.